+++ src/main/java/by/dero/gvh/model/itemsinfo/WebThrowInfo.java
@@ -5,7 +5,7 @@ import by.dero.gvh.model.ItemInfo;
 public class WebThrowInfo extends ItemInfo {
     private float multiplier;
     private int duration;
-    private int force;
+    private float force;
 
     public int getDuration() {
         return duration;
@@ -15,11 +15,11 @@ public class WebThrowInfo extends ItemInfo {
         this.duration = duration;
     }
 
-    public int getForce() {
+    public float getForce() {
         return force;
     }
 
-    public void setForce(int force) {
+    public void setForce(float force) {
         this.force = force;
     }
